+++ CHANGELOG.md
@@ -1,5 +1,47 @@
-# 1.0.0 (2023-02-03)
+# [1.3.0](https://github.com/swgoh-utils/swgoh-ae2/compare/v1.2.0...v1.3.0) (2024-04-08)
 
+
+### Bug Fixes
+
+* add AssetOS to Console ([fa93d09](https://github.com/swgoh-utils/swgoh-ae2/commit/fa93d098cd858b15272e7fe3874ed61cd2c87b53))
+* Update Dockerfile ([ebd5d67](https://github.com/swgoh-utils/swgoh-ae2/commit/ebd5d67745af36e47472a377f3d3b50d75bb573a))
+
+
+### Features
+
+* Allow Windows and set it to default ([7204712](https://github.com/swgoh-utils/swgoh-ae2/commit/720471238434a907128229011b7545746f939099))
+
+# [1.2.0](https://github.com/swgoh-utils/swgoh-ae2/compare/v1.1.1...v1.2.0) (2023-02-25)
+
+
+### Bug Fixes
+
+* ignore the GUI on build ([7b20794](https://github.com/swgoh-utils/swgoh-ae2/commit/7b20794bd6a1c3511eca6b14e1a7694eae4c96b2))
+
+
+### Features
+
+* added GUI (Windows only) ([d8c4d1f](https://github.com/swgoh-utils/swgoh-ae2/commit/d8c4d1ffe58e85ec4a64bb6c001da426058c0ae8))
+
+## [1.1.1](https://github.com/swgoh-utils/swgoh-ae2/compare/v1.1.0...v1.1.1) (2023-02-07)
+
+
+### Bug Fixes
+
+* add clear Description to enums. ([b1346fa](https://github.com/swgoh-utils/swgoh-ae2/commit/b1346fad2cec5083d5a0d2359795ceff77df4a89))
+
+# [1.1.0](https://github.com/swgoh-utils/swgoh-ae2/compare/v1.0.0...v1.1.0) (2023-02-06)
+
+
+### Features
+
+* Allow diffing versions ([311e084](https://github.com/swgoh-utils/swgoh-ae2/commit/311e084f7f11e90e000f8f2be3c6bb229fd8ec12))
+
+# 1.0.0 (2023-02-03)
+initial create
+
+# 1.0.1 (2023-02-06)
+added the ability to diff list
 
 ### Features